Purges any merged recipes from the PR.
This uses a `git` trick to find the recipes that are in `master` and
remove them. As this also picks up `example`, we no longer need to
handle that in a special way. The goal here is to avoid building or
checking things we don't need to. Further it is to improve the user
experience when the propose new additions. In particular, this should be
very helpful for new users that are frequently confused by this
behavior.