Kludge to work around strange OS X builder failure.
As written in a long, new comment:

The requirement of configparser below [in requirements.build] is a total hack.
For some reason, conda-build-all errors out when it tries to build multiple
versions of this package on the OSX builders: the first build succeeds, but
then it doesn't install pkg-config for the second build attempt for some
reason, leading to a failure. But, the only reason it was attempting to build
multiple versions of the package was because I was requiring `python`, which
is needed by X.org build tools but not actually used by the resulting binaries
-- so there should not be multiple output packages for different Python
versions. It turns out that conda-build-all decides to build multiple packages
if `python` is explicitly required in meta.yaml -- but not if it's implicitly
required. So, by requiring `configparser`, we draw in the Python dep needed to
do the build, but fool conda-build-all's logic, so that we only require one
package build and avoid the missing-pkg-config problem. Fun times.