Add Python 3.6 to matrix on AppVeyor
This explicitly adds an entry for Python 3.6 on AppVeyor and restricts
the Python 3.5 step to just Python 3.5 (not Python 3.5 and Python 3.6).
As we are merely shuffling the Python 3.6 builds into their own matrix
cases instead of having them lumped in with Python 3.5, this doesn't
actually result in any significant change in build time. Instead this
simply makes it easier for people to view and debug their builds on
AppVeyor at staged-recipes.

[skip lint]