Drop Python 3.4 AppVeyor update hack
This hack was added because Python 3.4's copy of `conda` was so old that
activation did not work properly. So instead of activating it right
away, this hack updated `conda` to a more recent version and then
activated. However, we have decided to drop Python 3.4. Further the
Python 3.5 Miniconda install on AppVeyor has been upgraded. Therefore a
hack of this nature is not really needed any more and can be dropped.

[skip lint]