Fix Homebrew uninstall step to ignore dependencies
Apparently `--force` is no longer sufficient when telling Homebrew to
uninstall packages. If it turns out the packages are require by other
packages, then `brew` will balk and issue an error message suggesting
one also use `--ignore-dependencies`. So we add it here in the hopes
that this will get Homebrew to comply with our uninstall request.

[skip lint]