xorg-xproto: quasi-hack to build on Windows 64
The 32-bit builds were OK, but config.guess wasn't able to figure out what OS the
Windows 64 builds were running on! The script also reported that it was from 2006, so
I have a hunch that a newer version will be smarter. One suspects that the GCC
config.guess will stay up-to-date.