xorg-xproto: reassuring Windows error!
A new error cropped up because our updated config.guess script changed the way
in which the host triple was reported, which broke a workaround in configure
that helped Windows build. Which implies that some effort was put in to make
these libraries build on Windows!

Although this is still getting close to not being worth the effort, since I
don't have any plans to build Windows versions of anything I use that needs
X11. But the OSX CI takes so long to run that I've got plenty of time to
iterate on Windows.

Also add a few args to configure that I think I will enjoy having in the
future.