xorg: Windows: transform key environment variables
Windows builds delegate to the build.sh script, but that meant that build.sh was
getting environment variables like $PREFIX with Windows-style paths. Problems
if they were ever echoed, etc. Transform the key paths in bld.bat, then. We also
add the -x option when running bash on the build script so we can trace it better.