Fast finish old Travis builds
Using Travis CI's API and information about the current build, this
script checks to see if the build number for this build is the most
recent one. If it doesn't have the most recent build number, it exits
with an error ending the build. However, if it does have the most recent
build number, then the build proceeds without issues.