Travis: Cache conda packages
To cutdown setup time on Travis CI, we implement caching for all `conda`
packages. As a result, anything used to setup the `root` environment or
anything used as part of the build will be cached. As PRs have separate
caches from the `master` branch on the repo, we don't have to worry
about lots of clutter collecting from different builds. Instead they
remain rather contained. If any cache gets messed up, it is a simple
matter to clear it and retry. This should reduce time spent downloading
packages that are reused often and hopefully make up for the absence of
a preincluded Miniconda.

[skip lint]
[skip appveyor]