Travis: Cache the Miniconda installer
The Miniconda installer itself is on the order of ~100MB. Skipping the
download of it each time and simply installing it from cache should help
speed up setup time. Thus we change this to download the Miniconda
installer directly into the cache and run it from there. If the
installer is not already present, we will download a new one to cache.
Thus if it is too old or people wish to freshen up the cache, this easy
to do through the CI GUI.

[skip lint]
[skip appveyor]