Hide requirements on Windows 32-bit
Something strange is happening with our `conda`-based tools, which is
causing them to look for a Windows 32-bit package even though Windows is
skipped entirely. Until we can sort it out, we can simply skip having
requirements show up on Window 32-bit. While a bit hacky, this appears
to get the job done and should allow the R stack to progress in
conda-forge while we try to sort out the underlying issue.

[skip appveyor]