Download fast finish script with PowerShell
It appears `curl` has moved off of the path on AppVeyor. While it is
possible to move `curl` back onto the path, there are a couple of
concerns. First this may result in including other things on the path
that we don't intend to. Second those other things may get involved in
builds corrupting packages we are trying to deploy. Third there was
never any guarantee that `curl` would be included on Windows nor is
there any guarantee it will still be there. So this changes the download
command to make use of PowerShell to perform the download. This is
something we can rely on. Therefore PowerShell should be a safe choice
for long term usage.

[skip lint]