.CI/create_feestocks.py: don't error out if rerendering produces no changes
By default Git will exit with an error if you try to make an empty commit.
This behavior could cause feedstock registration to abort if the `conda smithy
rerender` step had no changes to make. But in our system that's really not an
error, so let's use Git's `--allow-empty` option to signal that.

[ci skip]