Use @conda-forge-coordinator 's token temporarily
This creates a token under @conda-forge-coordinator as a temporary
workaround for creating repos at staged-recipes. As it is using the same
bot that is used to run team updates and feedstock updates, this will
get rate limited periodically. However it is better than not being able
to convert anything at all. Hopefully this can tide us over until we can
get this setup properly with @conda-forge-manager again.

[skip ci]