Bringing the OOMMFC package to conda-forge
The OOMMF Calculator (OOMMFC) is a calculator for the numerical
operations required for micromagnetic calculations. Technically,
the OOMMFC class inherits from the "micromagneticmodel", which is
an abstract description of a micromagnetic problem.

Another way of describing OOMMFC is that it provides a Python
interface to the OOMMF micromagnetic package from NIST.

This is the second last package from the Jupyter-OOMMF project
(http://joommf.github.io)