change tarball url (pypi to gitlab)
The .egg-info folder present in the tarball on pypi
seems to cause an error when building the package.
This is a workaround, but in next release we should change
the url back to pypi and ensure that this folder won't be
included, e.g., by using `git clean -xfd` on a specific
clone of the repository that is not used for local development
before uploading to pypi.