Move CircleCI configuration file for CircleCI 2.0
There are still many more changes needed beyond moving the configuration
file to this new location (and renaming it). However to make the commit
history and associated diffs easier to reason about, we simply move the
CircleCI configuration file in this commit. All the additional changes
for CircleCI 2.0 will be added in the following commit.

[skip lint]