Drop additions of maintainers
This is now handled through a web service, which takes care of team
updates for all feedstocks. The service picks up any changes pushed to
the feedstock and checks to see if any new members were added. If so, it
goes through adds them to the correct team. If not, nothing happens.
When it comes to a recently added feedstock, this web service will also
pick up the changes pushed to the recipe, create the team, and add all
members listed in recipe to it. Hence there is no need to maintain this
functionality separately in staged-recipes.

[ci skip]