Use conda-smithy's team addition feature
As a failsafe, have `conda-smithy` take over team additions for the
feedstocks. While it is true that team additions can be handled by the
webservice, it has the potential to have issues. As such, go ahead and
use `conda-smithy` to add teams first when creating the feedstock. This
ends up being no more code in `staged-recipes` and still allows us to
remove the duplicated team addition code that was here. Also make sure
to add teams only after pushing the final change to the feedstock.

[ci skip]