On Travis CI use Linux for pushes and Mac for PRs
Select which operating system to use for the VM based on whether the
build is a PR build or not. For PRs, use macOS as we are presumably
trying to build a recipe. For everything else, this should be a
feedstock conversion job. So using Linux in this case should be fine.
This uses Travis CI's new conditional matrix syntax to select the OS.
Hopefully can process conversions a bit faster.

ref: https://docs.travis-ci.com/user/conditional-builds-stages-jobs/

[skip lint]