Retrigger `staged-recipes` build from `script`
Seems that for some reason the `after_failure` section is not being run
when `script` errors out. To workaround this issue, try calling the
script to retrigger Travis CI in `script` after feedstock conversion
fails. Ensure that the build is still broken when this script is called.
Hopefully this way we can be sure that Travis CI is actually triggered
when feedstock creation fails.

[ci skip]