Pin M4RI
There has not been a M4RI release in years…I also don't know what @malb's plans
are for this. Since he's using date-of-release versioning, there is no clear
answer here. The next release could be compatible or not but the version number
does not help much here. So, in doubt, it'd say we do strict pinning.