Protect against `grep` match failure in `yum` step
When searching for `yum_requirements.txt`, it is possible to not find
any and thus not have anything for `grep` to parse. In this case, `grep`
will error out with exit code 1. This wasn't an issue before as we used
`grep` in a pipe. However now that `pipefail` is set, this no longer is
acceptable. To fix this, we suppress failures by `grep`. This allows
forwarding to `yum` to continue to work as expected without failing due
to an absence of `yum_requirements.txt`.

[skip lint]