Update meta.yaml
After talking to someone on the conda build team, they said that if I'm doing binary repackaging, I shouldn't be using binary relocation anyway.  They said I should just add the shared libraries that are assumed to be installed already to a whitelist.

Most of these are part of the manylinux1 specification anyway, and some shared libraries don't ever get loaded since this is a binary repackaging of a subsystem that not all of the parts are used for this client.  I may be able to weed more out in the future, but I'd like to get past the first step here.