Drop explicit `CPPFLAGS` and `LDFLAGS`
We specify `CPPFLAGS` and `LDFLAGS` in the compilers, which is currently
getting stomped on by these settings. Not to mention we include
`-headerpad_max_install_names` on macOS, which is needed to fix the
`install_name_tool` error that the build sees on macOS. Not to mention
all the flags being passed here are things we already include in the
compilers. So there is no need to specify them again here.