Force use of .NET's MSBuild on Visual Studio 2008
Borrows a nasty hack that was discovered a long time ago to ensure an
older `msbuild` is not picked up by accident. Forces the newer `msbuild`
from .NET to be picked up and used instead. This is needed for building
on Visual Studio 2008 where the activation script doesn't get this right
for some reason.