Workaround incomplete renaming of gdrive on GitHub
This was moved from a user repo to and org. However there are still some
old references to the name of the user repo. These seem to cause some
issues with the build as Go looks for the old repo name even though it
is not there. Here we add a hacky symlink to the old GitHub repo under
the user so that Go can find it.