Do not build on Windows
While the zfp documentation says that it can build on Windows, I cannot get it to do so.  I am no expert on Visual Studio, nor am I an expert on Windows, at all.  So, I'm punting.  I'd love it if someone else could figure out how to build zfp on the CI services that are used by conda-forge, namely VS9 (2008) on AppVeyor and VS14 (2015) on Azure Pipelines (if I understand it correctly).

Note to the future, according to the zfp documentation, only VS14 is actually supported, so I would not be surprised if the VS9 build failed.

Anyway, we can move forward with this build _now_ on MacOS and Linux, and I think we should.  If someone has the time and inclination to help us out with your Windows magic powers, we would be happy to put out a new build.