Add ECMWF magics-metview
A feedstock for magics already exists, but most users do not need
metview support. Magics when built with Metview support is much larger
and more complex. We can split it into a separate feedstock to make
things simpler for most users.

It's likely that only the metview-feedstock will use magics-metview.