Setting PKG_CONFIG_PATH is not necessary in epoxy build.sh.
pkg-config from conda-forge handles finding .pc files in CDT sysroot.
Incidentally, pkg-config fails to find gl and egl anyway, but the
libraries are only used for tests (disabled) and adding gl and egl to
requires.private in the epoxy.pc file. The following additional packages
would need to be installed so that pkg-config could find gl and egl:
    - {{ cdt('libdrm-devel') }}  # [linux]
    - {{ cdt('libxdamage-devel') }}  # [linux]
    - {{ cdt('libxext-devel') }}  # [linux]
    - {{ cdt('libxfixes-devel') }}  # [linux]
    - {{ cdt('libxxf86vm-devel') }}  # [linux]
    - libxcb  # [linux]
If that is done, then they would also need to be added to the gtk3
recipe so that it can find epoxy through pkg-config. Considering none of
that is needed for the package to work, I'll leave it out.