fix dependencies for cmake build
expat is a dependency of libcifti which should not be included in the
package.

some cmake versions on conda have some behavior that prevents the
cmake public_header property from being honored. 3.17 onwards does
not seem to have it but for now, staying on a low version of cmake
seems reasonable.

help2man is now used to generate man pages