Euphonic recipe: build against tweaked branch
Some of the build/compiler issues are due to setup.py making too many
assumptions. Hopefully we can improve compatibility with conda build
by no longer clobbering the CC env variable.

This is a temporary fix while build.yaml is being developed; once it's
working we can push a release to PyPI and build against that.