Attempt to pip-install semgrep wheel
Per https://github.com/returntocorp/semgrep/issues/2929 the semgrep
wheel on PyPI contains the binaries needed.

Alternatives to this approach would be to provide the URL for the wheel
directly in meta.yaml per

https://docs.conda.io/projects/conda-build/en/latest/user-guide/wheel-files.html

but I'm unable to find an easy non-hash-based URL for the wheels on the
PyPI website. This might be a good backup if the current approach in
this commit proves untenable.

Yet another alternative might be to build semgrep: ocaml and pkg-config
are both conda packages.