Build universal dummy wheel for Windows
Even though the rustboard binary now builds sucessfully on Windows, the
tensorboard-data-server build system explicitly states that Windows is
unsupported and cancels the build. I don't know if it could actually
work if I disable the check but I can't test if it works, so I will wait
for official support instead.