Add recipe for findorb
findorb (or Find_Orb) is a package for orbit determination and ephemeris
computation.  Find_Orb can take a set of observations of an asteroid, comet,
or natural or artificial satellite, given in the MPC (Minor Planet Center)
format, the newer ADES astrometric format, and/or the NEODyS or AstDyS
formats, and find the corresponding orbit.