WIP: attempt to build ancient 2.3.2 vde2 from sourceforge
peeled the onion a bit to see if I could get it to build.

There were straightforward patches for the first few issues but now
I'm hitting something that isn't fixed in the upstream (because they moved
to wolfssl) and I don't feel like continuing to work to make this build.

Code needs to fix allocation of EVP_CIPHER_CTX because the
type became opaque, and you can't allocate ctx on the stack any more.
The function to create a EVP_CIPHER_CTX *ctx is EVP_CIPHER_CTX_new to
move beyond openssl 1.1 and the rest of conda-forge is way past that (as it should be)