Explicitly add LICENSE.md for nmslibR
I confirmed that LICENSE.md is included in the tarball when
building from source. However, it's not included in the source
tarballs on CRAN for versions 1.0.6 or 1.0.5. IIRC I don't think
you are allowed to include a file named LICENSE.md. Instead of
getting involved in CRAN policies, I just added the file to the
recipe