rm project-provided CMAKE_CXX_STANDARD
systemc defaults to '98', so unless you hardcode a newer std c++98 or gnu++98
will be used.  By removing the default '98' from the systemc CMakeLists.txt
cmake will detect that the toolchain has an ABI baked into it and decide
not to use a -std option.  That's what we want, the toolchains builtin default.