feat: create `google-cloud-cpp-core` feedstock
`google-cloud-cpp` is too large to compile in a single feedstock. The CI
builds timeout before all the code is compiled.

This PR proposes a new `google-cloud-cpp-core` feedstock for
`google-cloud-cpp` that packages (1) the common libraries used by all
`google-cloud-cpp` features, and (2) the most popular features of
`google-cloud-cpp`.

These packages are small enough to compile in a single feedstock.

Future PRs will introduce other feedstocks that package more features of
`google-cloud-cpp`.

In these new feedstocks the runtime packages are named
`libgoogle-cloud-{feat}`, and the development packages are named
`libgoogle-cloud-{feat}-devel`.

The `libgoogle-cloud` / `libgoogle-cloud-devel` package pair contains
the common libraries used by all (or nearly all) subpackages.

There are over 100 libraries in `google-cloud-cpp`, most of them are
small, and most of them see little use. It is not worthwhile to create
packages for all these libraries. In a future PR we will introduce the
`libgoogle-cloud-all` and `libgoogle-cloud-all-devel` packages as catch
alls for the smaller and less used features.