feat: create `google-cloud-cpp-ai` feedstock
This feedstock will contain the AI and ML related features in
`google-cloud-cpp`. Some of the largest features in `google-cloud-cpp`
fall in this category, and the overall count of libraries is manageable.

Splitting these features to a separate feedstock will reduce the build
time for the `google-cloud-cpp` feedstock, and should keep it below the
CI timeouts in the medium term.