avoid target_platform is None in template conditional
I'm trying to avoid some of the preprocessor nastyness of this recipe by
using conditionals in a couple of places.  I'm being reminded that there is
a rendering pass that doesn't have the platform variant yaml is added in to
define the target_platform.  Seeing if we just make the meta.yaml evaluate
without blowing up for this first phase render if it works.  I vaguely
remember someone in core team suggesting this on another recipe I worked on
in the past.  Can't remember which one right now.