Remove pip check
The pip check is not required since this package is not expected to be installed by pip, only by conda. This is for two reasons: (1) it requires the non-Python dependency 'slim' from conda-forge; and (2) the dependency `msprime` has a broken recipe in which the 'newick' dependency is available from pip but not conda-forge.