Add ollama Python package
Note that ollama is already a package, so I've named this ollama-python (same
as the official git repo).

In principle, this could depend on ollama itself, but that package is quite
out of date. It isn't actually necessary to install ollama via conda, so for
now, I am leaving it out as a dependency.