Update recipes/hpv-kite/build.sh
From conda-forge pull request: danielnachun: Make careful note of the fact that `${PKG_VERSION}` is *not* escaped out so that it gets substituted in when `tee` is run. Also note that the shell and batch wrappers are installed on all platforms, but this is not a problem, because Unix users will just use the bash wrapper and Windows will only recognize the `.cmd` file as a valid executable.

Co-authored-by: Daniel Nachun <daniel.nachun@gmail.com>