Fix Windows build by copying os_support.h to source directory
The Windows build was failing because the vec.h file includes os_support.h,
but this file doesn't exist in the original RNNoise source. The os_support.h
file is provided in the recipe directory but needs to be copied to the src/
directory where the build expects to find it.

Fixes build error: 'Cannot open include file: os_support.h': No such file or directory